% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{calculateAllPositions}
\alias{calculateAllPositions}
\title{Calculate Positions Relative to an Arbitrary Anchor (Vectorized)}
\usage{
calculateAllPositions(result, mw, allele, anchor_offset = 0L)
}
\arguments{
\item{result}{A GRanges object containing the results of a motif scan, with
metadata columns for window index, context start and end, and motif strand
for both alleles. The metadata columns should be named in the format
"windowIdxRef", "contextStartRef", "contextEndRef", "motifStrandRef" for
the reference allele and similarly for the alternate allele.}

\item{mw}{Integer (or vector). Motif width.}

\item{allele}{Character. "Ref" or "Alt" to specify which allele's context
coordinates to use.}

\item{anchor_offset}{Integer. The offset from the context start to use as the
anchor point (0-based).}
}
\value{
A data.frame containing absolute and relative coordinates for all
  variants.
}
\description{
Calculate Positions Relative to an Arbitrary Anchor (Vectorized)
}
