% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{calculatePvalue}
\alias{calculatePvalue}
\title{Calculate the significance of the matches for the reference and alternate alleles for their PWM}
\source{
Hélène Touzet and Jean-Stéphane Varré (2007) Efficient and accurate P-value computation for Position Weight Matrices.
 Algorithms for Molecular Biology, \bold{2: 15}.
}
\usage{
calculatePvalue(
  results,
  granularity = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{results}{The output of \code{motifbreakR} that was run with \code{filterp=TRUE}}

\item{granularity}{Numeric; the granularity to which to round the PWM,
larger values compromise full accuracy for speed of calculation. A value of
\code{NULL} does no rounding.}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam-class}} object
controlling parallel evaluation. Try \code{BiocParallel::registered()} to
see what is available; the default \code{BiocParallel::SerialParam()} gives
serial evaluation.}
}
\value{
a GRanges object. The same GRanges object that was input as \code{results}, but with
 \code{pValueRef} and \code{pValueAlt} columns in the output modified from \code{NA} to the p-value
 calculated by \code{\link[TFMPvalue]{TFMsc2pv}}. Additionally a \code{pValueEffect} column that indicates "strong"
 when the two p-values differ by more than an order of magnitude, otherwise "weak".
}
\description{
Calculate the significance of the matches for the reference and alternate alleles for their PWM
}
\details{
This function is intended to be used on a selection of results produced by \code{\link{motifbreakR}}, and
this can be (although not always) a very memory and time intensive process if the algorithm doesn't converge rapidly.
The nucleotide background used is the one stored by \code{motifbreakR} in
\code{attributes(results)$bkg}.
}
\examples{
data(example.results)
rs1006140 <- example.results[example.results$SNP_id \%in\% "rs1006140"]
# low granularity for speed; 1e-6 or 1e-7 recommended for accuracy
rs1006140 <- calculatePvalue(rs1006140, BPPARAM=BiocParallel::SerialParam(), granularity = 1e-4)


}
\seealso{
See \code{\link[TFMPvalue]{TFMsc2pv}} from the \pkg{TFMPvalue} package for
  information about how the p-values are calculated.
}
