% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.results}
\alias{example.results}
\title{Example Results from motifbreakR}
\format{
\code{\link[GenomicRanges]{GRanges}} output from \code{motifbreakR}
}
\usage{
data(example.results)
}
\value{
\code{\link[GenomicRanges]{GRanges}} object. See \code{\link{motifbreakR}} for information on it's structure.
}
\description{
This contains example results from motifbreaker for use in examples from the help docs
}
\examples{
data(example.results)
example.results

}
\keyword{datasets}
