% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{exportMBtable}
\alias{exportMBtable}
\title{Export motifbreakR results to csv or tsv}
\usage{
exportMBtable(results, file, format = "tsv")
}
\arguments{
\item{results}{The output of \code{motifbreakR}}

\item{file}{Character; the file name of the destination file}

\item{format}{Character; one of tsv (tab separated values) or csv (comma separated values)}
}
\value{
\code{exportMBtable} produces an output file containing the output
of the \code{motifbreakR} function.
}
\description{
Export motifbreakR results to csv or tsv
}
\examples{
data(example.results)
example.results
\donttest{
exportMBtable(example.results, file = tempfile(fileext = ".tsv"), format = "tsv")
}
}
\seealso{
See \code{\link{exportMBbed}} for the function that exports the
\code{motifbreakR} results as a BED file, colored by selected score.
}
