% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{scoreAllWindows}
\alias{scoreAllWindows}
\title{scoreMotif.a <- function(snp.seq, ppm, len, offset = 1) {
  snp.seq <- snp.seq[offset:(offset + len - 1)]
  ## diag code
  position.probs <- c(ppm[snp.seq, ])[1L + 0L:(len - 1L) * (len + 1L)]
  return(position.probs)
}
#' @importFrom compiler cmpfun
scoreMotif <- cmpfun(scoreMotif.a, options = list(optimize = 3))}
\usage{
scoreAllWindows(snp.seq, snp.seq.rc, pwm, from = "default",
  to = "default", pwm.range = NULL, calcp = TRUE)
}
\description{
scoreMotif.a <- function(snp.seq, ppm, len, offset = 1) {
  snp.seq <- snp.seq[offset:(offset + len - 1)]
  ## diag code
  position.probs <- c(ppm[snp.seq, ])[1L + 0L:(len - 1L) * (len + 1L)]
  return(position.probs)
}
#' @importFrom compiler cmpfun
scoreMotif <- cmpfun(scoreMotif.a, options = list(optimize = 3))
}
