% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{shiny_motifbreakR}
\alias{shiny_motifbreakR}
\title{Run Shiny version of the motifbreakR package}
\usage{
shiny_motifbreakR()
}
\value{
returns a \code{\link[shiny]{shinyAppDir}} that launches the shiny app when printed.
}
\description{
Run Shiny version of the motifbreakR package
}
\examples{
library(motifbreakR)

app <- shiny_motifbreakR()

if (interactive()) {
  shiny::runApp(app)
}
}
