% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scoreMotif.R
\name{motifbreakR}
\alias{motifbreakR}
\title{Predict The Disruptiveness Of Single Nucleotide Polymorphisms On
Transcription Factor Binding Sites.}
\usage{
motifbreakR(snpList, pwmList, threshold, method = "default", bkg = NULL,
  show.neutral = FALSE, verbose = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{snpList}{The output of \code{snps.from.rsid} or \code{snps.from.bed}}

\item{pwmList}{An object of class \code{MotifList} containing the motifs that
you wish to interrogate}

\item{threshold}{Numeric; the minimum disruptiveness score for which to
report results}

\item{method}{Character; one of \code{default}, \code{log}, or \code{IC}; see
details.}

\item{bkg}{Numeric Vector; the background probabilites of the nucleotides
used with method=\code{log} method=\code{IC}}

\item{show.neutral}{Logical; include neutral changes in the output}

\item{verbose}{Logical; if running serially, show verbose messages}

\item{BPPARAM}{a BiocParallel object see \code{\link{[BiocParallel]register}}
and see \code{getClass("BiocParallelParam")} for additional parameter
classes}
}
\value{
a GRanges object containing:
 \item{REF}{the reference allele for the SNP}
 \item{ALT}{the alternate allele for the SNP}
 \item{snpPos}{the coordinates of the SNP}
 \item{motifPos}{the coordinates of the SNP within the TF binding motif}
 \item{geneSymbol}{the geneSymbol corresponding to the TF of the TF binding motif}
 \item{dataSource}{the source of the TF binding motif}
 \item{providerName, providerId}{the name and id provided by the source}
 \item{seqMatch}{the sequence on the 5' -> 3' direction that corresponds to DNA
 at the position that the TF binding motif was found.}
 \item{pctRef}:
 \item{pctAlt}:
 \item{alleleRef}:
 \item{alleleAlt}:
 \item{effect}{one of weak, strong, or neutral indicating the strength of the effect.}
}
\description{
Predict The Disruptiveness Of Single Nucleotide Polymorphisms On
Transcription Factor Binding Sites.
}
\details{
\pkg{motifbreakR} works with position probability matrices (PPM).
  PPM are derived as the fractional occurrence of nucleotides A,C,G, and T at
  each position of a position frequency matrix (PFM). PFM are simply the
  tally of each nucleotide at each position across a set of aligned
  sequences. With a PPM, one can generate probabilities based on the genome,
  or more practically, create any number of position specific scoring
  matrices (PSSM) based on the principle that the PPM contains information
  about the likelihood of observing a particular nucleotide at a particular
  position of a true transcription factor binding site. What follows is a
  discussion of the three different algorithms that may be employed in calls
  to the \pkg{motifbreakR} function via the \code{method} argument. Suppose
  we have a PPM \eqn{P} with columns \eqn{i} and \eqn{j} rows for nucleotides
  A,C,G,T. We can produce scores from the matrix corresponding to a sequence
  at position \eqn{n}. Suppose further that the vector of scores \eqn{p}
  derived from \eqn{P} given the sequence is \eqn{p_n}. The sum of this
  vector is called the match score\cite{pssm}, \eqn{S}: \deqn{S(p_n)~=~\sum
  p_{n}} When weighted by the difference between the best nucleotide at each
  position of \eqn{P} minus the worst, we have a weighted-sum, or the default
  method of motif scoring in motifbreakR. This algorithm may be represented
  mathematically as \deqn{S(p_n)~=~p_{n}~\cdot~\omega_{P}} where \eqn{\omega}
  is the \eqn{max(P_{i}) - min(P_{i})} for each column \eqn{i} of the
  probability matrix. This is the method of \citet{plosgen77}. The classic
  PWM is defined as the sum of the log-probabilities, accessed from
  method="log": \deqn{S(p_n)~=~\sum {log (p_n/p_{bk})}} where the background
  probabilities of the letters are represented by \eqn{p_{bk}}. By default a
  uniform distribution is assumed, but the user may specify any distribution
  by including the "bkg" argument in function calls. This parameter is a
  vector (length 2) specifying the probability of A and G in the genome as a
  fraction of 1. Lastly, the scores may be weighted by self-information
  content (IC), or relative entropy, at each position of the scoring
  matrix\cite{Stormo2000}. The method, accessed by specifying method="IC"
  (information content) is philosophically similar to the weighted matrix-sum
  above. \deqn{S(p_n)~=~p_{n}~\cdot~I_{P}} where the vector of information
  content at all positions of the matrix, \(I_P\), is given by
  \deqn{I_{P}~=~\sum_{j=A}^{T} P_{i,j}log_2 (P_{i,j}/p_{bk})} Regardless of
  the called method, all scores are normalized to the scoring range over the
  maximum and minimum values of $P$. Thus the final score \eqn{S_{norm}} is
  given as \deqn{S_{norm}~=~\frac{S(p_n) - S(P_{min})}{S(P_{max})}}
}
\seealso{
\code{\link{snps.from.rsid}} \code{\link{snps.from.bed}}
}

