% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus.from.rsid.R
\name{snps.from.file}
\alias{snps.from.file}
\alias{variants.from.file}
\title{Import variants from a BED file or VCF file for use in motifbreakR}
\usage{
snps.from.file(
  file = NULL,
  dbSNP = NULL,
  search.genome = NULL,
  format = "bed",
  biomart.dataset = NULL,
  check.unnamed.for.rsid = FALSE
)

variants.from.file(
  file = NULL,
  dbSNP = NULL,
  search.genome = NULL,
  biomart.dataset = NULL,
  format = "bed"
)
}
\arguments{
\item{file}{Character; the path to a BED or VCF file; see Details for the
required format}

\item{dbSNP}{OPTIONAL; an object of class SNPlocs to look up rsids; run
\code{\link[BSgenome]{available.SNPs}()} to list the available SNPlocs packages}

\item{search.genome}{an object of class BSgenome for the species you are interrogating;
see \code{\link[BSgenome]{available.genomes}} for a list of species}

\item{format}{Character; one of \code{bed} or \code{vcf}}

\item{biomart.dataset}{a Mart object from \code{\link[biomaRt]{useEnsembl}} specifying
the \code{snps} biomart, which dataset i.e., \code{hsapiens_snp}, and which
version i.e., \code{111} or \code{GRCm39}. This will override \code{SNPlocs} and must
be compatible with search.genome selection, and will query from \code{biomaRt},
which may be considerably faster than a lookup from a \code{SNPlocs} object.}

\item{check.unnamed.for.rsid}{Logical; for BED input, look up variants named
in the form chr:pos:ref:alt in dbSNP and use the rsid if one is found. Lookup
may be slow and requires either \code{dbSNP} or \code{biomart.dataset}.}
}
\value{
a GRanges object with one range per variant and alternate allele
 (multi-allelic variants are split into one range per alternate allele),
 containing:
 \item{SNP_id}{The identifier of the variant}
 \item{REF}{The reference allele, as a DNAStringSet}
 \item{ALT}{The alternate allele, as a DNAStringSet}
 The name of the BSgenome package is stored in \code{attributes(x)$genome.package}
 for use by \code{\link{motifbreakR}}.
}
\description{
Import variants from a BED file or VCF file for use in motifbreakR
}
\details{
\code{snps.from.file} reads a BED or VCF file and generates the input
 for \code{motifbreakR}. Both SNVs and indels are imported.

 For BED files see \url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}
 for a complete description of the format. Our convention deviates in that
 there is a required format for the name field: \code{name} is either
 chromosome:start:REF:ALT or an rsid from dbSNP (in which case \code{dbSNP}
 or \code{biomart.dataset} must be supplied). For example rs123 would be
 entered in the first format as chr7:24966446:C:A. Several alternate alleles
 may be separated by commas, e.g. chr7:24966446:C:A,G.

 For VCF files, multi-allelic records are split into one variant per
 alternate allele. Variants whose alleles contain characters outside the
 IUPAC DNA alphabet (for example symbolic alleles such as \code{<DEL>}) are
 removed with a warning.
}
\section{Functions}{
\itemize{
\item \code{variants.from.file()}: Identical to \code{snps.from.file}; retained for
backwards compatibility from when \code{snps.from.file} excluded indels by default.

}}
\examples{
 library(BSgenome.Drerio.UCSC.danRer7)
 snps.bed.file <- system.file("extdata", "danRer.bed", package = "motifbreakR")
 # see the contents
 read.table(snps.bed.file, header = FALSE)
 #import the BED file
 snps.mb <- snps.from.file(snps.bed.file,
                           search.genome = BSgenome.Drerio.UCSC.danRer7,
                           format = "bed")

}
\seealso{
See \code{\link{motifbreakR}} for analysis; See \code{\link{snps.from.rsid}}
  for an alternate method for generating a list of variants.
}
