% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus.from.rsid.R
\name{snps.from.rsid}
\alias{snps.from.rsid}
\title{Import SNPs from rsid for use in motifbreakR}
\usage{
snps.from.rsid(
  rsid = NULL,
  dbSNP = NULL,
  search.genome = NULL,
  biomart.dataset = NULL
)
}
\arguments{
\item{rsid}{Character; a character vector of rsid values from dbSNP}

\item{dbSNP}{an object of class SNPlocs to look up rsids; run
\code{\link[BSgenome]{available.SNPs}()} to list the available SNPlocs packages}

\item{search.genome}{an object of class BSgenome for the species you are interrogating;
see \code{\link[BSgenome]{available.genomes}} for a list of species.}

\item{biomart.dataset}{a Mart object from \code{\link[biomaRt]{useEnsembl}} specifying
the \code{snps} biomart, which dataset i.e., \code{hsapiens_snp}, and which
version i.e., \code{111} or \code{GRCm39}. This will override \code{SNPlocs} and must
be compatible with search.genome selection, and will query from \code{biomaRt},
which may be considerably faster than a lookup from a \code{SNPlocs} object.}
}
\value{
a GRanges object with one range per variant and alternate allele
 (multi-allelic variants are split into one range per alternate allele),
 containing:
 \item{SNP_id}{The identifier of the variant}
 \item{REF}{The reference allele, as a DNAStringSet}
 \item{ALT}{The alternate allele, as a DNAStringSet}
 The name of the BSgenome package is stored in \code{attributes(x)$genome.package}
 for use by \code{\link{motifbreakR}}.
}
\description{
Import SNPs from rsid for use in motifbreakR
}
\details{
\code{snps.from.rsid} takes an rsid, or a character vector of rsids,
 and generates the input required by \code{motifbreakR}. rsids that are not
 found are dropped with a warning.
}
\examples{
 library(BSgenome.Hsapiens.UCSC.hg19)
 library(SNPlocs.Hsapiens.dbSNP155.GRCh37)
 snps.file <- system.file("extdata", "pca.enhancer.snps", package = "motifbreakR")
 snps <- as.character(read.table(snps.file)[,1])
 snps.mb <- snps.from.rsid(snps[1],
                           dbSNP = SNPlocs.Hsapiens.dbSNP155.GRCh37,
                           search.genome = BSgenome.Hsapiens.UCSC.hg19)

 ## alternatively using biomaRt

 library(biomaRt)
 library(BSgenome.Hsapiens.UCSC.hg38)
 ensembl_snp <- useEnsembl(biomart = "snps",
                           dataset = "hsapiens_snp",
                           version = "112")
 snps.mb <- snps.from.rsid(snps,
                           biomart.dataset = ensembl_snp,
                           search.genome = BSgenome.Hsapiens.UCSC.hg38)

}
\seealso{
See \code{\link{motifbreakR}} for analysis; See \code{\link{snps.from.file}}
  for an alternate method for generating a list of variants.
}
